+++ pkg/server/ingest.go
@@ -146,14 +146,16 @@ func (h ingestHandler) ingestInputFromRequest(r *http.Request) (*ingestion.Inges
 		}
 
 	case format == "pprof":
-		input.Profile = &pprof.RawProfile{RawData: b}
 		input.Format = ingestion.FormatPprof
+		input.Profile = &pprof.RawProfile{
+			RawData: b,
+		}
+
 	case strings.Contains(contentType, "multipart/form-data"):
 		input.Profile = &pprof.RawProfile{
 			FormDataContentType: contentType,
 			RawData:             b,
 		}
-		input.Format = ingestion.FormatPprof
 	}
 
 	if input.Profile == nil {